+++ src/components/Hotspots/Variants/DataHotspot.svelte
