+++ src/components/Hotspots/Variants/InfoHotspot.svelte
