+++ src/components/Hotspots/Variants/SceneHotspot.svelte
